feat(utils): Add support for CPM model and corresponding functionalities.
This commit introduces support for the CPM model in the ImgHandler class. It includes loading and using the CPM model and tokenizer, as well as handling image and text processing for the CPM model. New methods have been added to accommodate these changes, enabling the extraction of features from images and texts through the CPM model.