feat(scene_detection): add grid_size parameter to frame extraction
This commit introduces a new `grid_size` parameter to the `extract_frames_with_cv` function within the `SceneDetector` class. The addition of this parameter allows users to specify the grid size for frame extraction, providing greater flexibility in scene analysis. By default, the grid size is set to (3, 3), but it can now be customized when calling the `detect` method, enhancing the functionality of the scene detection process.